remove TestingLang; simplify to StateLang
