+++ testing_languages/blockchain_test_lang.py
@@ -2,16 +2,11 @@
 import re
 import random as r
 
-from testing_languages.testing_language import TestingLanguage
+from simulations.state_language import StateLanguage
 from casper.protocols.blockchain.blockchain_protocol import BlockchainProtocol
-from casper.networks import NoDelayNetwork
-from casper.plot_tool import PlotTool
-from casper.safety_oracles.clique_oracle import CliqueOracle
-from casper.validator_set import ValidatorSet
-import casper.utils as utils
 
 
-class BlockchainTestLang(TestingLanguage):
+class BlockchainTestLang(StateLanguage):
     """Allows testing of simulation scenarios with small testing language."""
 
     # Signal to py.test that TestLangCBC should not be discovered.
@@ -22,7 +17,6 @@ class BlockchainTestLang(TestingLanguage):
 
     def check_estimate(self, validator, estimate):
         """Check that a validators forkchoice is some block"""
-        self.check_validator_exists(validator)
         self.check_message_exists(estimate)
 
         message = self.messages[estimate]
@@ -34,7 +28,6 @@ class BlockchainTestLang(TestingLanguage):
 
     def check_safe(self, validator, estimate):
         """Check that some validator does not detect safety on a block."""
-        self.check_validator_exists(validator)
         self.check_message_exists(estimate)
 
         message = self.messages[estimate]
@@ -47,7 +40,6 @@ class BlockchainTestLang(TestingLanguage):
 
     def check_unsafe(self, validator, estimate):
         """Must be implemented by child class"""
-        self.check_validator_exists(validator)
         self.check_message_exists(estimate)
 
         message = self.messages[estimate]
