+++ tests/casper/langs/test_test_lang.py
@@ -1,7 +1,7 @@
 """The language testing module ... """
 import pytest
 
-from testing_languages.testing_language import TestingLanguage
+from testing_languages.blockchain_test_lang import BlockchainTestLang
 
 
 #TODO: add tests for TestingLanguage!
